Added Tables, Entities & basic validation

+++ src/Controller/UsersController.php
@@ -19,19 +19,38 @@ class UsersController extends AppController
         $this->Auth->allow(['sign', 'login', 'register', 'forgot', 'reset']);
     }
 
+    /**
+     * Register action
+     */
     public function sign () {
-        $this->loadModel('Cities');
 
-        $cities = $this->Cities->find('list', [
+        $user = null;
+
+        if ($this->request->is('post')) {
+            $regUser = TableRegistry::get('Users');
+            $user = $regUser->newEntity($this->request->getData());
+
+            if ($regUser->save($user)) {
+                $this->Flash->success(__("Inscription terminée avec succès ! Vous pouvez dès à présent vous connecter."));
+            } else {
+                $this->Flash->error(__("Des champs ne sont pas valides, veuillez corriger les erreurs"));
+            }
+        }
+
+        $cities = TableRegistry::get('Cities')->find('list', [
             'keyField' => 'id',
             'valueField' => 'name'
         ])->toArray();
 
         $cities = ['' => 'Ville'] + $cities;
 
-        $this->set(compact('cities'));
+        $this->set(compact('cities', 'user'));
     }
 
+    /**
+     * Sign in action /w 13 months cookie support
+     * @return \Cake\Http\Response|null
+     */
     public function login () {
         if ($this->request->is('post')) {
             $user = $this->Auth->identify();
@@ -54,31 +73,39 @@ class UsersController extends AppController
 
                 return $this->redirect($this->Auth->redirectUrl());
             } else {
-                $this->Flash->error(__('Les identifiants sont incorrectes'));
+                $this->Flash->error(__('Aucun utiliateur ne correspond à ces identifiants.'));
             }
             return $this->redirect($this->referer());
         }
     }
 
+    /**
+     * Logout action (destroy auth session/cookie)
+     */
     public function logout () {
-        TableRegistry::get('Tokens')->deleteAll([
-            'name' => 'auth_token',
-            'user_id' => $this->Auth->user('id')
-        ]);
-        $this->Cookie->delete('auth_token');
-        $this->Auth->logout();
+        if ($this->request->is('post')) {
+            TableRegistry::get('Tokens')->deleteAll([
+                'name' => 'auth_token',
+                'user_id' => $this->Auth->user('id')
+            ]);
+            $this->Cookie->delete('auth_token');
+            $this->Auth->logout();
+        } else {
+            $this->redirect('/');
+        }
     }
 
-    public function register () {
-        $regUser = TableRegistry::get('Users');
-        $user = $regUser->newEntity($this->request->getData());
-        // TODO: validation
-    }
-
+    /**
+     * Password request action
+     */
     public function forgot () {
 
     }
 
+    /**
+     * Password reset action
+     * @param $token
+     */
     public function reset ($token) {
 
     }
